# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is a sample API documentation.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/holiday": {
            "get": {
                "tags": [
                    "HolidayPlan"
                ],
                "summary": "Get list of holiday",
                "description": "Returns list of holiday",
                "operationId": "9e386d5c1c3b9cd729b7e01c39a73ce2",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HolidayPlan"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "HolidayPlan"
                ],
                "summary": "Create holiday",
                "description": "Create holiday",
                "operationId": "1dd5d65bd04b6be28980decdaae7656a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "startHoliday": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "endHoliday": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/holiday/{id}": {
            "get": {
                "tags": [
                    "HolidayPlan"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "276cf5b7f3ac33ac2d98c1eebd20edf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Holiday plan to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HolidayPlan"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "HolidayPlan"
                ],
                "summary": "Update plan",
                "description": "Update plan",
                "operationId": "d26ef5f14b4eb17fafdc445089ddf08c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Holiday plan to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "startHoliday": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "endHoliday": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "HolidayPlan"
                ],
                "summary": "Delete user",
                "description": "Delete users",
                "operationId": "4b99d2b8a6e3a71b1609d86d517d6435",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Holiday plan to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HolidayPlan"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/holidays/pdf": {
            "get": {
                "tags": [
                    "HolidayPlan"
                ],
                "summary": "Get pdf list of holiday",
                "description": "Returns list of holiday",
                "operationId": "805f2eb4b9828bae2908d5e85835a16c",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HolidayPlan"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/holidays/pdf/{id}": {
            "get": {
                "tags": [
                    "HolidayPlan"
                ],
                "summary": "Get pdf list of users",
                "description": "Returns pdf",
                "operationId": "b8f279cad0285abc7de9ee931a60aa09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Holiday plan to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HolidayPlan"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "9cb2efee96b203086a6d6e3a2dcff2b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "b1ce2afcdc7bf63047937cb7ea1141e3",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "7e8c77661442d3c113e6e8a7751ac813",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update users",
                "description": "Update users",
                "operationId": "60097a376b965f0ec6eac581b62cc61e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create users",
                "description": "Create users",
                "operationId": "71d86dd873f9e567415d5da5929a0dea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "Delete users",
                "operationId": "1bddc8611e319cd89fee27928ebf5345",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "HolidayPlan": {
                "properties": {
                    "id": {
                        "description": "Holiday ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Holiday plan, name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Holiday plan, description",
                        "type": "string"
                    },
                    "startHoliday": {
                        "description": "Start holiday plan",
                        "type": "string",
                        "format": "date"
                    },
                    "endHoliday": {
                        "description": "End holiday plan",
                        "type": "string",
                        "format": "date"
                    },
                    "location": {
                        "description": "Holiday plan, location",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string"
                    },
                    "password": {
                        "description": "User password",
                        "type": "string"
                    },
                    "password_confirmation": {
                        "description": "User password confirmation",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a sanctum token to authorize",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "HolidayPlan",
            "description": "HolidayPlan"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}